+++ incl/dofuture_OP.R
@@ -1,0 +1,5 @@
+y <- foreach(x = 1:10, .combine = rbind) %dofuture% {
+  y <- sqrt(x)
+  data.frame(x = x, y = y, pid = Sys.getpid())
+}
+print(y)